+++ src/user/user.service.ts
@@ -1,11 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable } from '@nestjs/common';
+import { REQUEST } from '@nestjs/core';
+import { Request } from 'express';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UserService {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    @Inject(REQUEST) private readonly request: Request,
+  ) {}
 
   async create(createUserDto: CreateUserDto) {
     return await this.prisma.user.create({ data: createUserDto });
@@ -20,10 +25,14 @@ export class UserService {
   }
 
   async findByUsername(username: string) {
-    return await this.prisma.user.findUnique({ where: { username: username } });
+    return await this.prisma.user.findUniqueOrThrow({
+      where: { username: username },
+    });
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const userId = this.request['user']['sub'];
+    if (userId !== id) throw new ForbiddenException();
     return await this.prisma.user.update({
       where: { id: id },
       data: updateUserDto,
@@ -31,6 +40,8 @@ export class UserService {
   }
 
   async remove(id: number) {
+    const userId = this.request['user']['sub'];
+    if (userId !== id) throw new ForbiddenException();
     return await this.prisma.user.delete({ where: { id: id } });
   }
 }
